requirements.txt
Flask==2.2.5
google-api-python-client==2.88.0
google-auth==2.21.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
requests==2.32.3
gunicorn==21.2.0
